# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NFL Survivor Pool API",
    "version": "1.0.0",
    "description": "API for NFL Survivor Pool Web App"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Entry": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "pool": {
            "type": "string"
          },
          "request": {
            "type": "string"
          },
          "entryNumber": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "eliminated"
            ]
          },
          "eliminatedWeek": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Pick": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "entry": {
            "type": "string"
          },
          "entryNumber": {
            "type": "integer"
          },
          "team": {
            "type": "string"
          },
          "week": {
            "type": "integer"
          },
          "game": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "away_team": {
                "type": "string"
              },
              "home_team": {
                "type": "string"
              },
              "event_date": {
                "type": "string",
                "format": "date-time"
              },
              "schedule": {
                "type": "object",
                "properties": {
                  "week": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      },
...
      "Pool": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "season": {
            "type": "integer"
          },
          "currentWeek": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "completed",
              "open"
            ]
          },
          "maxParticipants": {
            "type": "integer"
          },
          "entryFee": {
            "type": "number"
          },
          "prizeAmount": {
            "type": "number"
          },
          "creator": {
            "type": "string"
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "eliminatedUsers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "maxEntries": {
            "type": "integer"
          },
          "prizePot": {
            "type": "number"
          }
        }
      },
      "PoolInput": {
        "type": "object",
        "required": [
          "name",
          "season",
          "maxParticipants",
          "entryFee",
          "prizeAmount",
          "description",
          "startDate",
          "endDate",
          "maxEntries",
          "prizePot"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "season": {
            "type": "integer"
          },
          "maxParticipants": {
            "type": "integer"
          },
          "entryFee": {
            "type": "number"
          },
          "prizeAmount": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "maxEntries": {
            "type": "integer"
          },
          "prizePot": {
            "type": "number"
          }
        }
      },
      "PoolWithUserInfo": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/Pool"
          },
          {
            "type": "object",
            "properties": {
              "userRequests": {
                "type": "integer"
              },
              "userEntries": {
                "type": "integer"
              },
              "canJoin": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "PoolWithEntries": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/Pool"
          },
          {
            "type": "object",
            "properties": {
              "activeEntries": {
                "type": "integer"
              },
              "userEntryId": {
                "type": "string"
              }
            }
          }
        ]
      },
      "PoolStats": {
        "type": "object",
        "properties": {
          "totalParticipants": {
            "type": "integer"
          },
          "eliminatedParticipants": {
            "type": "integer"
          },
          "currentWeek": {
            "type": "integer"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized access",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/pools/{poolId}/entries": {
      "get": {
        "summary": "Get all entries for a pool",
        "tags": [
          "Entries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entry"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      }
    },
...
    "/api/v1/picks/pool/{poolId}": {
      "get": {
        "summary": "Get all picks for a pool",
        "tags": [
          "Picks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the pool"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pick"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/api/v1/pools": {
      "get": {
        "summary": "Get all pools",
        "tags": [
          "Pools"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "select",
            "schema": {
              "type": "string"
            },
            "description": "Fields to select (comma-separated)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Sort criteria (field:asc/desc)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of pools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pool"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new pool",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoolInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created pool",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Pool"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/pools/available": {
      "get": {
        "summary": "Get available pools for the current user",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of available pools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PoolWithUserInfo"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/v1/pools/user": {
      "get": {
        "summary": "Get pools for the current user",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's pools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PoolWithEntries"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/v1/pools/user/active": {
      "get": {
        "summary": "Get active pools for the current user",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's active pools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PoolWithEntries"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/v1/pools/{id}": {
      "get": {
        "summary": "Get a single pool",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pool details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Pool"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      },
      "put": {
        "summary": "Update a pool",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoolInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated pool",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Pool"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a pool",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pool deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      }
    },
    "/api/v1/pools/{id}/stats": {
      "get": {
        "summary": "Get pool statistics",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pool statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PoolStats"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      }
    },
    "/api/v1/pools/{id}/entries": {
      "get": {
        "summary": "Get entries for a pool",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of entries for the pool",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entry"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      }
    },
    "/api/v1/pools/{id}/status": {
      "put": {
        "summary": "Update pool status",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "open",
                      "closed",
                      "completed"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pool status updated",
            "content": {
              "application/json": {
                "schema": null,
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "data": {
                    "$ref": "#/components/schemas/Pool"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      }
    },
...
    "/api/v1/requests/pool/{poolId}": {
      "get": {
        "summary": "Get all requests for a specific pool (Admin only)",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pool requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Request"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
...
}